nvim 0.7.x: Fix missing parameter in vim.lsp.util.apply_text_edits
The function has a new "offset_encoding" parameter which is "utf-8",
"utf-16" or "utf-32". Because this parameter was missing, an error
message was shown every time one attempted to autocomplete a function
name by hitting enter.

It's not really clear what value to use here, but since LSP is based on
utf-16, I'm using utf-16 for now.